Add files via upload
This is my project. The programme reads .srt file (subtitles) and .txt file (learner's vocabulary / corpus). The programme replaces all words in the .srt file with *** (according to the number of letters) if these words are present in .txt file. The programme saves the edited .srt file. The practical use: helps in L2 teaching and learning, as it enhances listening practice and still allows to read difficult words while watching a film in L2.

P.S. I didn't use pipelines in this programme, because not one, but three files are involved and it was difficult for me to make it. Also I googled a lot, and used several functions we didn't cover during our course.